simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
]
